+++ apps/server/src/modules/oauth/uc/index.ts
@@ -1,1 +1,2 @@
 export * from './oauth.uc';
+export * from './hydraOauth.uc';
