+++ config/globals.js
@@ -106,6 +106,9 @@ const globals = {
 	NEXTCLOUD_BASE_URL: process.env.NEXTCLOUD_BASE_URL,
 	NEXTCLOUD_ADMIN_USER: process.env.NEXTCLOUD_ADMIN_USER,
 	NEXTCLOUD_ADMIN_PASS: process.env.NEXTCLOUD_ADMIN_PASS,
+	NEXTCLOUD_CLIENT_ID: process.env.NEXTCLOUD_CLIENT_ID,
+	NEXTCLOUD_CLIENT_SECRET: process.env.NEXTCLOUD_CLIENT_SECRET,
+	HYDRA_URI: process.env.HYDRA_URI,
 
 	// calendar
 	CALENDAR_URI: process.env.CALENDAR_URI,
